Repackaging: * Move source files into treescriptify_text package * Move test files into single test package * Move __main__.py into treescriptify_text package
